+++ include/dctl/zobrist/accumulate.hpp
@@ -2,7 +2,6 @@
 #include <boost/range/numeric.hpp>      // accumulate
 #include <array>                        // array
 #include <cassert>                      // assert
-#include <cstddef>                      // size_t
 
 namespace dctl {
 namespace zobrist {
@@ -12,7 +11,7 @@ auto accumulate(SinglePassRange const& rng, std::array<T, N> const& arr)
 {
         return boost::accumulate(
                 rng, T{0}, [&](auto const& hash, auto const& elem) {
-                return hash ^ arr[static_cast<std::size_t>(elem)];
+                return hash ^ arr[elem];
         });
 }
